fix: Correct outlet list functionality
This commit fixes a regression where the first outlet was only outputting the first bar of a completed span instead of the full list.

The issue was caused by a type mismatch in the call to `outlet_list`. The function expects the list size as a `short`, but a `long` was being passed. This has been corrected by explicitly casting the size argument to `(short)`.

Additionally, all other calls to `outlet_list` and `outlet_anything` have been updated with the correct cast to prevent similar issues.

+++ buildspans/buildspans.c
@@ -990,7 +990,7 @@ void buildspans_end_track_span(t_buildspans *x, t_symbol *track_sym) {
             outlet_int(x->track_outlet, track_num_to_output);
 
             // Outlet 1: Span list
-            outlet_list(x->span_outlet, NULL, span_size, span_atoms);
+            outlet_list(x->span_outlet, NULL, (short)span_size, span_atoms);
         }
 
         if (local_span_created) {
@@ -1207,7 +1207,7 @@ void buildspans_prune_span(t_buildspans *x, t_symbol *track_sym, long bar_to_kee
             outlet_int(x->track_outlet, track_num_to_output);
 
             // Outlet 1: Span list
-            outlet_list(x->span_outlet, NULL, span_size, span_atoms);
+            outlet_list(x->span_outlet, NULL, (short)span_size, span_atoms);
         }
         object_free(ended_span_array_for_validation);
     }
@@ -1607,15 +1607,15 @@ void buildspans_output_span_data(t_buildspans *x, t_symbol *track_sym, t_atomarr
                 if (type == A_FLOAT) {
                     t_atom out_atom;
                     atom_setfloat(&out_atom, atom_getfloat(&a));
-                    outlet_anything(x->log_outlet, output_key_sym, 1, &out_atom);
+                    outlet_anything(x->log_outlet, output_key_sym, (short)1, &out_atom);
                 } else if (type == A_LONG) {
                     t_atom out_atom;
                     atom_setlong(&out_atom, atom_getlong(&a));
-                    outlet_anything(x->log_outlet, output_key_sym, 1, &out_atom);
+                    outlet_anything(x->log_outlet, output_key_sym, (short)1, &out_atom);
                 } else if (type == A_SYM) {
                     t_atom out_atom;
                     atom_setsym(&out_atom, atom_getsym(&a));
-                    outlet_anything(x->log_outlet, output_key_sym, 1, &out_atom);
+                    outlet_anything(x->log_outlet, output_key_sym, (short)1, &out_atom);
                 } else if (type == A_OBJ) {
                     t_object *obj = atom_getobj(&a);
                     if (object_classname(obj) == gensym("atomarray")) {
@@ -1623,7 +1623,7 @@ void buildspans_output_span_data(t_buildspans *x, t_symbol *track_sym, t_atomarr
                         long ac;
                         t_atom *av;
                         atomarray_getatoms(arr, &ac, &av);
-                        outlet_anything(x->log_outlet, output_key_sym, ac, av);
+                        outlet_anything(x->log_outlet, output_key_sym, (short)ac, av);
                     }
                 }
             }
